Make EarlyStopper use floats

+++ src/models/early_stopper.py
@@ -1,12 +1,12 @@
 class EarlyStopper:
     def __init__(self, patience: int = 10, min_delta: int = 0):
-        self.min_metric: int = 0
+        self.min_metric: float = 0.0
         self.count: int = 0
         self.patience = patience
         self.min_delta = min_delta
         self.stop = False
 
-    def check_early_stop(self, metric: int):
+    def check_early_stop(self, metric: float):
         if metric < self.min_metric:
             self.min_metric = metric
             self.count = 0
